Breaking the PFI examples into different files in a subfolder.
